Validações das entradas de dados

+++ DeleteProduct/index.js
@@ -5,17 +5,31 @@ const { ObjectID } = require('mongodb')
 module.exports = async function (context, req) {
     const { id } = req.params;
 
+    if (!id) {
+        context.res = {
+            status: 400,
+            body: 'Provide a product id on params',
+        };
+        return;
+    }
+
     const { client: MongoClient, closeConnectionFn } = await createMongoClient();
     const Products = MongoClient.collection('products');
 
-    const deletedProduct = await Products.findOneAndDelete({ _id: ObjectID(id) });
+    try {
+        const deletedProduct = await Products.findOneAndDelete({ _id: ObjectID(id) });
+        closeConnectionFn();
 
-    closeConnectionFn();
+        context.res = {
+            status: 200,
+            body: 'Product deleted successfully'
+        };
 
-    context.res = {
-        status: 200,
-        body: deletedProduct
-    };
-
+    } catch (error) {
+        context.res = {
+            status: 500,
+            body: 'Error on delete product ' + id,
+        };
+    }
 
 }
